# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CMS manager",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost/cms-template/public",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/customer/orders": {
            "get": {
                "tags": [
                    "Customer Orders"
                ],
                "summary": "Get all orders for the authenticated customer",
                "operationId": "a937ee7c3a9e2429f1d853976b2f933d",
                "responses": {
                    "200": {
                        "description": "List of customer's orders"
                    }
                },
                "security": [
                    {
                        "customer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customer Orders"
                ],
                "summary": "Create an order from the customer's cart",
                "operationId": "f1bd737905588e5425a3cad730a83df3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "shipping_address_id"
                                ],
                                "properties": {
                                    "shipping_address_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Leave at the front door"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created from cart"
                    },
                    "400": {
                        "description": "Cart is empty or invalid input"
                    }
                },
                "security": [
                    {
                        "customer": []
                    }
                ]
            }
        },
        "/api/customer/orders/update-from-cart/{orderId}": {
            "put": {
                "tags": [
                    "Customer Orders"
                ],
                "summary": "Update an order to match the customer's current cart",
                "operationId": "6d319061f8f605c0394b8e58490acd2f",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order updated from cart"
                    },
                    "400": {
                        "description": "Cart is empty or update failed"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "customer": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Authentication using Bearer token (Sanctum)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "cookieSession": {
                "type": "apiKey",
                "description": "Authentication using session cookie (laravel_session)",
                "name": "laravel_session",
                "in": "cookie"
            },
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
...
}